move cmd to internal, remove init methods
