+++ internal/cmd/local/prerun.go
@@ -2,47 +2,42 @@ package local
 
 import (
 	"fmt"
-	"github.com/airbytehq/abctl/internal/local/k8s"
-	"github.com/airbytehq/abctl/internal/telemetry"
-	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
-	"os"
-	"path/filepath"
 )
 
-// persistentPreRunLocal is extracted from the cobra.Command initialization for reading purposes only
-func persistentPreRunLocal(cmd *cobra.Command, _ []string) error {
-	// telemetry client configuration
-	{
-		// ignore the error as it will default to false if an error returns
-		dnt, _ := cmd.Flags().GetBool("dnt")
-		var telOpts []telemetry.GetOption
-		if dnt {
-			telOpts = append(telOpts, telemetry.WithDnt())
-		}
+//// persistentPreRunLocal is extracted from the cobra.Command initialization for reading purposes only
+//func persistentPreRunLocal(cmd *cobra.Command, _ []string) error {
+//	// telemetry client configuration
+//	{
+//		// ignore the error as it will default to false if an error returns
+//		dnt, _ := cmd.Flags().GetBool("dnt")
+//		var telOpts []telemetry.GetOption
+//		if dnt {
+//			telOpts = append(telOpts, telemetry.WithDnt())
+//		}
+//
+//		telClient = telemetry.Get(telOpts...)
+//	}
+//	// provider configuration
+//	{
+//		var err error
+//		provider, err = k8s.ProviderFromString(flagProvider)
+//		if err != nil {
+//			return err
+//		}
+//
+//		printK8sProvider(provider)
+//	}
+//
+//	return nil
+//}
 
-		telClient = telemetry.Get(telOpts...)
-	}
-	// provider configuration
-	{
-		var err error
-		provider, err = k8s.ProviderFromString(flagProvider)
-		if err != nil {
-			return err
-		}
-
-		printK8sProvider(provider)
-	}
-
-	return nil
-}
-
-func printK8sProvider(p k8s.Provider) {
-	userHome, _ := os.UserHomeDir()
-	configPath := filepath.Join(userHome, p.Kubeconfig)
-	pterm.Info.Printfln("using kubernetes provider:\n  provider name: %s\n  kubeconfig: %s\n  context: %s",
-		p.Name, configPath, p.Context)
-}
+//func printK8sProvider(p k8s.Provider) {
+//	userHome, _ := os.UserHomeDir()
+//	configPath := filepath.Join(userHome, p.Kubeconfig)
+//	pterm.Info.Printfln("using kubernetes provider:\n  provider name: %s\n  kubeconfig: %s\n  context: %s",
+//		p.Name, configPath, p.Context)
+//}
 
 // preRunInstall is extracted from the cobra.Command initialization for reading purposes only
 func preRunInstall(cmd *cobra.Command, _ []string) error {
